+++ src/main/java/com/springboot/blog/service/impl/ToaNhaServiceImpl.java
@@ -3,6 +3,7 @@ package com.springboot.blog.service.impl;
 import com.springboot.blog.entity.CanHo;
 import com.springboot.blog.entity.LoaiCanHo;
 import com.springboot.blog.entity.ToaNha;
+import com.springboot.blog.exception.BlogAPIException;
 import com.springboot.blog.exception.ResourceNotFoundException;
 import com.springboot.blog.payload.CanHoDto;
 import com.springboot.blog.payload.ToaNhaDto;
@@ -11,6 +12,7 @@ import com.springboot.blog.repository.ToaNhaRepository;
 import com.springboot.blog.service.ToaNhaService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,12 +22,14 @@ import java.util.stream.Collectors;
 public class ToaNhaServiceImpl implements ToaNhaService {
 
     private ToaNhaRepository toaNhaRepository;
+    private CanHoRepository canHoRepository;
 
     private ModelMapper modelMapper;
     @Autowired
-    public ToaNhaServiceImpl(ToaNhaRepository toaNhaRepository, ModelMapper modelMapper) {
+    public ToaNhaServiceImpl(ToaNhaRepository toaNhaRepository, ModelMapper modelMapper, CanHoRepository canHoRepository) {
         this.toaNhaRepository = toaNhaRepository;
-        this.modelMapper = modelMapper; // Tiêm modelMapper vào
+        this.modelMapper = modelMapper;
+        this.canHoRepository = canHoRepository;
     }
     @Override
     public ToaNhaDto addToaNha(ToaNhaDto toaNhaDto) {
@@ -62,8 +66,9 @@ public class ToaNhaServiceImpl implements ToaNhaService {
     public void deleteToaNha(Long toaNhaId) {
         ToaNha toaNha = toaNhaRepository.findById(toaNhaId)
                 .orElseThrow(()-> new ResourceNotFoundException("Toa nha", "id", toaNhaId));
-
-        toaNhaRepository.delete(toaNha);
+        List<CanHo> canHo = canHoRepository.findByToaNhaId(toaNhaId);
+        if(canHo == null) toaNhaRepository.delete(toaNha);
+        else throw new BlogAPIException(HttpStatus.NOT_ACCEPTABLE,"khong the xoa toa nha");
 
     }
 }
